+++ ThirdPartyAdapters/pubmatic/pubmatic/src/test/kotlin/com/google/ads/mediation/pubmatic/PubMaticNativeAdTests.kt
@@ -96,6 +96,7 @@ class PubMaticNativeAdTests {
         mediationAdLoadCallback,
         pubMaticAdFactory,
         Dispatchers.Unconfined,
+        isRtb = true,
       )
       .onSuccess { pubMaticNativeAd = it }
   }